Tests/LibWeb/Text/expected/HTML/HTMLMediaElement-readyState-progression-audio.txt

loadstart: readyState >= HAVE_NOTHING
durationchange: readyState >= HAVE_METADATA duration=12.0675
loadedmetadata: readyState >= HAVE_METADATA
loadeddata: readyState >= HAVE_CURRENT_DATA
canplay: readyState >= HAVE_FUTURE_DATA
canplaythrough: readyState >= HAVE_ENOUGH_DATA
seeking: readyState >= HAVE_METADATA
seeked: readyState >= HAVE_METADATA
